chore(web): rename file
